+++ db/migrate/20140609212406_create_topics.rb
@@ -1,0 +1,13 @@
+class CreateTopics < ActiveRecord::Migration
+  def change
+    create_table :topics do |t|
+      t.string :name
+      t.integer :user_id
+      t.boolean :stickied, null: false, default: false
+      t.boolean :locked, null: false, default: false
+      t.boolean :deleted, null: false, default: false
+
+      t.timestamps
+    end
+  end
+end