+++ db/migrate/20140609214314_create_comments.rb
@@ -1,0 +1,12 @@
+class CreateComments < ActiveRecord::Migration
+  def change
+    create_table :comments do |t|
+      t.integer :topic_id
+      t.text :description
+      t.integer :user_id
+      t.boolean :deleted, null: false, default: false
+
+      t.timestamps
+    end
+  end
+end